libhwc: Fix for directFB out-of-buf access.
This patch is required to solve the fb driver PAGE fault issue.
There is a corner case in FIMD direct FB mechanism.
FIMD dma_end address will be out of the mapped region for the following scenario.
When ((src_buf_left  != 0) || (disp_left < 0))  &&
    ((src_buf_top + (disp_top < 0 ? disp_top : 0 ) + disp_h) == src_vstride)
This patch is a workaround to resolve the issue.

Change-Id: Ibb88414e8db9254876d5e7bd15592a331ee8f31b
Signed-off-by: meka rama <v.meka@samsung.com>