libhwcService: Grouped S3D functions under their own macro
S3D is no longer a subset of HWC_SERVICE so we need a new macro

Also added a check to prevent resolution change when the new dimensions
are equal to the current ones.
This is now necessary to stop the TV from unnecessarily blinking when
playing S3D videos.

Change-Id: I7c52bcd076b9322dcf0d2e9a8fd9cbb8634abe35
Signed-off-by: Jason Lee <aznair.lee@samsung.com>