libion_exynos: Patch for avoiding the zero fd from ion_alloc
Zero fd is a valid number in Linux. But, it could cause a problem
in many libraries for error handling.

Usually zero fd is reserved for stdin; ion_alloc could get the
zero fd when stdin is freed. Sometimes, kernel is failed to import
the zero-fd buffer.
To avoid any such side effects, this patch ensures that ion-alloc
always returns a non zero fd.

Change-Id: I72c09d30d5b765aa496fe810a7edfe6543672eeb
Signed-off-by: meka rama <v.meka@samsung.com>